PMW3901, PAA5100: bindings into single module for #301.
