--- micropython/modules/breakout_paa5100/micropython.cmake
@@ -1,20 +1,0 @@
-set(MOD_NAME breakout_paa5100)
-string(TOUPPER ${MOD_NAME} MOD_NAME_UPPER)
-add_library(usermod_${MOD_NAME} INTERFACE)
-
-target_sources(usermod_${MOD_NAME} INTERFACE
-    ${CMAKE_CURRENT_LIST_DIR}/${MOD_NAME}.c
-    ${CMAKE_CURRENT_LIST_DIR}/${MOD_NAME}.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/../../../libraries/${MOD_NAME}/${MOD_NAME}.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/../../../drivers/pmw3901/pmw3901.cpp
-)
-
-target_include_directories(usermod_${MOD_NAME} INTERFACE
-    ${CMAKE_CURRENT_LIST_DIR}
-)
-
-target_compile_definitions(usermod_${MOD_NAME} INTERFACE
-    -DMODULE_${MOD_NAME_UPPER}_ENABLED=1
-)
-
-target_link_libraries(usermod INTERFACE usermod_${MOD_NAME})
