+++ micropython/modules/micropython.cmake
@@ -16,7 +16,6 @@ include(breakout_roundlcd/micropython)
 include(breakout_rgbmatrix5x5/micropython)
 include(breakout_matrix11x7/micropython)
 include(breakout_msa301/micropython)
-include(breakout_paa5100/micropython)
 include(breakout_pmw3901/micropython)
 include(breakout_mics6814/micropython)
 include(breakout_potentiometer/micropython)
